rename the module so that it won't conflict the function name
