+++ tensorflow/contrib/nn/__init__.py
@@ -28,7 +28,7 @@ from __future__ import print_function
 # pylint: disable=unused-import,wildcard-import
 from tensorflow.contrib.nn.python.ops.cross_entropy import *
 from tensorflow.contrib.nn.python.ops.sampling_ops import *
-from tensorflow.contrib.nn.python.ops.modrelu import *
+from tensorflow.contrib.nn.python.ops.mod_relu import *
 # pylint: enable=unused-import,wildcard-import
 
 from tensorflow.python.util.all_util import remove_undocumented
