+++ tensorflow/contrib/nn/python/ops/mod_relu.py
